chore: minor code cleanup
